+++ CHANGELOG.md
@@ -8,7 +8,7 @@
     -   form type **slider**: replaced deprecated `terra-slider` by a `mat-slider` variant.
     -   form type **checkbox group**: replaced deprecated `tc-checkbox-group` by a `mat-select` variant in the multiple mode.
     -   form type **color**: replaced deprecated `terra-color-picker` by a `mat-input` variant.
-    -   form type **datepicker**: replaced deprecated `terra-date-picker` by a `mat-datepicker` variant.
+    -   form type **date**: replaced deprecated `terra-date-picker` by a `mat-datepicker` variant.
     -   form type **text**: replaced deprecated `terra-text-input` by a `mat-input` variant.
     -   form type **checkbox**: replaced deprecated `terra-checkbox` by a `mat-checkbox` variant.
     -   form type **select**: replaced deprecated `terra-select-box` by a `mat-select` variant.